feat(ui): add UI log control functions - Added functions to control the visibility of UI logs. - Updated the main view to include a toggle for UI logs.
